refactor(FRONTEND): :recycle: mejorada logica de filtrado para buscar en objetos anidados
