+++ client/src/components/TablaTickets.jsx
@@ -58,7 +58,7 @@ export const TablaTickets = ({ items }) => {
   ))
 
   return (
-    <section>
+    <>
       <Table hover bordered striped className='text-center text-uppercase'>
         <thead className='bg-primary text-white'>
           <tr>
@@ -69,6 +69,6 @@ export const TablaTickets = ({ items }) => {
         <tbody className='bg-secondary text-primary'>{renderItems}</tbody>
       </Table>
       <div className='d-flex justify-content-center'>{renderPagination}</div>
-    </section>
+    </>
   )
 }
